Edit `build-constants.go` to hard-code bad values
When a consumer does not run `gow generate`, e.g. if they are using
`go get` to attempt to pull the root dependency, without ever generating
these constants, it is made clearer at runtime that there is an issue.

Adding and modifying these values from development machine is done in
- a5d1df1aec7e6ad8e26e939202bb0e821b31abc5
- ffcf789a0bf3093706ed668cd12d296bccdd464a
- 9420f7a3a71b11136ed82d51d5b49169b565780c
and should never be done again.

Fixes: #3
Related-to: a5d1df1aec7e6ad8e26e939202bb0e821b31abc5
Signed-off-by: Mahied Maruf <contact@mechite.com>